order_service/pyproject.toml
[tool.poetry]
name = "order-service"
version = "0.1.0"
description = ""
authors = ["Illia Troshchynskyi <itroshchinskiy@rambler.ua>"]
readme = "README.md"
packages = [{include = "order_service"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.101.1"}
python-dotenv = "^1.0.0"
beanie = "^1.21.0"
asgi-lifespan = "2.1.0"
polyfactory = "2.7.2"
pydantic = "^2.2.0"
pydantic-settings = "^2.0.3"
pytest-coverage = "^0.0"
httpx = "^0.24.1"
prometheus-fastapi-instrumentator = "^6.1.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
flake8-print = "^5.0.0"
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
pytest-httpx = "^0.22.0"
flake8-quotes = "^3.3.2"
flake8-import-order = "^0.18.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
